First steps of splitting API logically. Users ops are now detached from main API
